hms_backend/requirements.txt
Django>=5.0
psycopg2-binary
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
